+++ app/src/main/java/com/ooredoo/bizstore/ui/activities/HomeActivity.java
@@ -307,8 +307,7 @@ public class HomeActivity extends AppCompatActivity implements
         mSearchResultsListView.invalidateViews();
     }
 
-    public void showDetailActivity(int detailType, String dealCategory, long typeId)
-    {
+    public void showDetailActivity(int detailType, String dealCategory, int typeId) {
         Intent intent = new Intent();
         intent.setClass(this, detailType == BUSINESS ? BusinessDetailActivity.class : DealDetailActivity.class);
         intent.putExtra(AppConstant.ID, typeId);
